+++ src/components/Common/UserContext/UserProviderWrapper.tsx
@@ -1,0 +1,20 @@
+import { useMemo, useState } from 'react';
+import { UserProvider } from '.';
+
+export default function UserProviderWrapper({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  const [loginId, setLoginId] = useState<number | null>(null);
+
+  const contextValue = useMemo(
+    () => ({
+      loginId,
+      setLoginId,
+    }),
+    [loginId, setLoginId],
+  );
+
+  return <UserProvider {...contextValue}>{children}</UserProvider>;
+}
